charter: Remove voting member categories
The BSC has been considering the role of the voting member categories.
These categories may provide positive reinforcement of a range of views
from different aspects of the eBPF ecosystem. At the same time, they can
be restrictive in determining membership of the committee such that
certain voices that do not fit the category system are missing.

Given the eligibility for these categories currently, it is difficult to
define a list of nominees for the upcoming election. This amendment
proposes to eliminate the categories entirely in order to allow for a
wider range of views, and defer the decision to the electorate.

This proposal is derived from a prior proposal by Lisa Caywood.

Signed-off-by: Joe Stringer <joe@cilium.io>