charter: Expand technical project scope
During the 2023-04-19 BSC meeting, members of the BSC expressed
interest in directing funds towards open source projects that are
essential to the eBPF ecosystem, which are not governed by the eBPF
foundation and likely never can be, such as the Linux kernel or LLVM.
Furthermore, some additional work such as security audits or preparation
of training material may be useful for driving the mission of the eBPF
foundation and may or may not be tied specifically to projects under the
eBPF foundation.

This proposal expands the meaning of a technical project to be either
the existing meaning (an open source project submitted under the eBPF
foundation), or a new meaning: a technical work item that is expected to
be scoped out in coordination with the BSC. It is expected to be backed
by a statement of work with details such as cost, duration, expected
outcome and so on.

Signed-off-by: Joe Stringer <joe@cilium.io>